requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
crispy-bootstrap5==0.6
Django==4.0.6
django-crispy-forms==1.14.0
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
lxml==4.9.1
oauthlib==3.2.0
protobuf==4.21.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.11
